libp2p: Refactor unittests
Refactor libp2p network unittests to integration tests.
For this, the network has an extra configuration flag to enable the
memory transport. This enables the usage of the memory transport
outside of the network unittests.
